Extract OwnProfilePage and OthersProfilePage from ProfilePage
No functional changes, just trying to get filesize slightly down.
